Make H1SpecExceptions a class
